// src/app/modules/reactive-forms/src/lib/components/rf-select-date-picker/models/rf-select-date-picker-months.model.ts
export interface RfSelectDatePickerMonths {
  january: string;
  february: string;
  march: string;
  april: string;
  may: string;
  june: string;
  july: string;
  august: string;
  september: string;
  october: string;
  november: string;
  december: string;
}

export interface RfSelectDatePickerOptionsData {
  day: string;
  month: string;
  year: string;
  months: RfSelectDatePickerMonths;
}